// App/src/types/storage.d.ts
export interface storage_data_MK {
    No: number;
    MK: string;
    SKS: number;
    Sem: number;
    Prasyarat: string;
    Dosen: string;
    Jadwal: string
}

export interface storage_selected_MK {
    kode: number;
    senin: number;
    selasa: number;
    rabu: number;
    kamis: number;
    jumat: number;
    sabtu: number;
}